Fix proto parsing failures
There were times when reading from cache that the proto parsing would
fail. What was happening was that we were writing the proto and the new
line character in two separate write() calls. If the proto was large
enough, then the first write() call would flush the write buffer to disk
along with the proto. The second write() call with the newline would
then be buffered and have to wait until flushed. In the read side, we
are assuming that the readline() would return a line with the last
character being a newline char. Because the second write() call was
buffered, we would read a line without a newline character failing the
parse. The fix is to simply concat the newline character when writing
the proto to file.

Change-Id: Id47ded346c2c083df5b9e652bbd5cd9c9506f867